move LayerItem to LayerManager
move LayerItem closer to LayerManager, to reflect the relation
of LayerItem as nested child to LayerManager
