--- src/modules/map/components/layerItem/index.js
@@ -1,4 +1,0 @@
-import { LayerItem } from './LayerItem';
-if (!window.customElements.get(LayerItem.tag)) {
-	window.customElements.define(LayerItem.tag, LayerItem);
-}
