+++ src/modules/map/components/layerManager/LayerItem.js
